add tests for restructured functions
