+++ rising/transforms/format.py
@@ -18,6 +18,8 @@ class MapToSeq(AbstractTransform):
             additional keyword arguments passed to superclass
         """
         super().__init__(grad=grad, **kwargs)
+        if isinstance(keys[0], (list, tuple)):
+            keys = keys[0]
         self.keys = keys
 
     def forward(self, **data) -> tuple:
