+++ tests/transforms/test_format_transforms.py
@@ -8,14 +8,14 @@ class TestFormat(unittest.TestCase):
         torch.manual_seed(0)
 
     def test_map_to_seq(self):
-        trafo = MapToSeq('data', 'seg', 'label')
+        trafo = MapToSeq(('data', 'seg', 'label'))
         out = trafo(**{'data': 0, 'seg': 1, 'label': 2})
         self.assertEqual(out[0], 0)
         self.assertEqual(out[1], 1)
         self.assertEqual(out[2], 2)
 
     def test_seq_to_map(self):
-        trafo = SeqToMap('data', 'seg', 'label')
+        trafo = SeqToMap(('data', 'seg', 'label'))
         out = trafo(0, 1, 2)
         self.assertEqual(out['data'], 0)
         self.assertEqual(out['seg'], 1)
